# api-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bitexroom api",
        "description": "Description of your API",
        "contact": {
            "email": "mehdints@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/2fa/setup": {
            "get": {
                "tags": [
                    "Two-Factor Authentication"
                ],
                "summary": "Setup Two-Factor Authentication",
                "description": "Generates a QR code and secret key for setting up two-factor authentication (2FA).",
                "operationId": "setup2FA",
                "responses": {
                    "200": {
                        "description": "QR code and secret key for setting up 2FA",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TwoFactorSetupResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. User is not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/2fa/save-secret": {
            "post": {
                "tags": [
                    "Two-Factor Authentication"
                ],
                "summary": "Save Two-Factor Authentication Secret",
                "description": "Validates and saves the 2FA secret key provided by the user.",
                "operationId": "save2FASecret",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "2fa",
                                    "secret"
                                ],
                                "properties": {
                                    "2fa": {
                                        "description": "The 2FA token entered by the user.",
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "secret": {
                                        "description": "The secret key used for generating 2FA tokens.",
                                        "type": "string",
                                        "example": "JBSWY3DPEHPK3PXP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Secret saved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Two-factor authentication secret saved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error or invalid 2FA token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid 2FA token or secret."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/2fa/verify-login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Validate two-factor authentication code and generate access token",
                "description": "This endpoint accepts a two-factor authentication (2FA) code from the user and validates it. If the code is correct, it generates and returns an access token that can be used to authenticate further requests. The API will reject any invalid or missing 2FA code with an error.",
                "operationId": "validateTwoFactor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ValidateTwoFactorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Two-factor validation successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "$ref": "#/components/schemas/AccessTokenResource"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidateTwoFactorRequest": {
                "required": [
                    "google2fa"
                ],
                "properties": {
                    "google2fa": {
                        "description": "The user's two-factor authentication code.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "AccessTokenResource": {
                "properties": {
                    "access_token": {
                        "description": "The generated access token.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TwoFactorSetupResponse": {
                "title": "TwoFactorSetupResponse",
                "description": "Response schema for 2FA setup",
                "properties": {
                    "qr_image": {
                        "description": "The URL of the QR code image for setting up 2FA.",
                        "type": "string",
                        "format": "uri"
                    },
                    "secret_key": {
                        "description": "The secret key for setting up 2FA.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}